+++ src/models/parser.model.ts
@@ -34,6 +34,7 @@ export interface ParsedUserConfigurationFile {
 
 export interface ParsedUserConfiguration {
     configurationTitle: string,
+    parserId: string,
     imageProviders: string[],
     steamDirectory: string,
     appendArgsToExecutable: boolean,
@@ -83,9 +84,9 @@ export type NameVariables = 'EXENAME' | 'FILENAME';
 export type ExtensionVariables = 'EXEEXT' | 'FILEEXT';
 export type PathVariables = 'EXEPATH' | 'FILEPATH';
 export type ParserVariables = 'TITLE' | 'FUZZYTITLE' | 'FINALTITLE';
-export type OtherVariables = '/';
+export type EnvironmentVariables = '/' | 'SRMDIR';
 
-export type AllVariables = DirectoryVariables | NameVariables | ExtensionVariables | PathVariables | ParserVariables | OtherVariables;
+export type AllVariables = DirectoryVariables | NameVariables | ExtensionVariables | PathVariables | ParserVariables | EnvironmentVariables;
 
 export interface GenericParser {
     getParserInfo(): ParserInfo,