+++ src/renderer/schemas/app-settings.schema.ts
@@ -25,6 +25,7 @@ export const appSettings = {
             default: {},
             properties: {
                 retrieveCurrentSteamImages: { type: 'boolean', default: true },
+                deleteDisabledShortcuts: { type: 'boolean', default: true },
                 imageZoomPercentage: { type: 'number', default: 40, minimum: 30, maximum: 100 },
                 preload: { type: 'boolean', default: false },
             }